fix handleAuctionStarted
