+++ schema.graphql
@@ -1,19 +1,3 @@
-# type Account @entity {
-#     "Account address"
-#     id: ID!
-#     wallet: String!
-#     balance: BigInt!
-#     historicalBalances: [HistoricalBalance!] @derivedFrom(field: "account")
-# }
-
-type HistoricalBalance @entity {
-    id: ID!
-    account: Account!
-    balance: BigInt!
-    timestamp: DateTime!
-}
-
-
 type Auction @entity {
   id: ID! #id is a required field
   blockNum: Int!
