+++ src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './Balances'
 export * from './Registrar'
 export * from './Slots'
 export * from './Auctions'
